# Optimize value conversions back and forth between driver.Value and concrete types.