### Description
Due to issues with the design of the driver API (see https://github.com/golang/go/issues/6918) there are dynamic memory allocations that happen when converting back and forth between the concrete types and `driver.Value`.
The change presented here allows the application to work around this situation by passing values by pointer to pre-allocated values of the base types supported by the database/sql package.

Here's a profile showing the issue:
<img width="219" alt="screen shot 2017-06-10 at 2 05 25 pm" src="https://user-images.githubusercontent.com/865510/27006320-ef1defbe-4de5-11e7-9567-a351076f14c9.png">

Here are some performance numbers before and after (the benchmark is from a private repo, I can craft a benchmark in go-sql-driver/mysql to highlight the problem if you wish):

```
BenchmarkJobDB/open_a_transaction,_put_jobs,_commit,_and_dispose-4         	     200	  66371357 ns/op	 1531183 B/op	   20208 allocs/op
PASS
ok  	github.com/segmentio/centrifuge/mysql	20.728s
```
```
BenchmarkJobDB/open_a_transaction,_put_jobs,_commit,_and_dispose-4         	     200	  64491960 ns/op	 1162803 B/op	    4209 allocs/op
PASS
ok  	github.com/segmentio/centrifuge/mysql	19.421s
```
```
benchmark                                                              old ns/op     new ns/op     delta
BenchmarkJobDB/open_a_transaction,_put_jobs,_commit,_and_dispose-4     66371357      64491960      -2.83%

benchmark                                                              old allocs     new allocs     delta
BenchmarkJobDB/open_a_transaction,_put_jobs,_commit,_and_dispose-4     20208          4209           -79.17%

benchmark                                                              old bytes     new bytes     delta
BenchmarkJobDB/open_a_transaction,_put_jobs,_commit,_and_dispose-4     1531183       1162803       -24.06%
```

What this is doing is making Go believe that an interface that holds a `*int64` is holding a value to a `int64` (and similar for other types) so [this check](https://golang.org/src/database/sql/convert.go#L113) passes in the database/sql package, then once the value comes back into the driver, it converts it back to the actual type it's supposed to have.

Honestly I'm a bit sad/ashamed of having to hack the Go type system to work around this issue, but the gains are so massive that I believe the change deserves to be at least discussed.

Because this is likely non-portable and may break in future versions of Go I think we should make it an opt-in features enabled by a runtime flag.

Please take a look and let me know what you think.

### Checklist
- [x] Code compiles correctly
- [ ] Created tests which fail without the change (if possible)
- [x] All tests passing
- [ ] Extended the README / documentation, if necessary
- [x] Added myself / the copyright holder to the AUTHORS file