@achille-roussel 2017-06-15
I've opened an issue for a more general solution to the efficiency problem => https://github.com/go-sql-driver/mysql/issues/620
It seems to me like it would be better than introducing the hack that's in this PR, let me know.