@arnehormann 2017-06-24
We will not allow unsafe anywhere in the driver, it is a constraint in some contexts and the internals useable by it are not guaranteed to be stable.