@achille-roussel 2017-06-24
Fair enough, it did solve the issue I had but I wasn't really happy with the implementation either. 

Thanks for taking a look!